Support macro loading per file.
This patch provides a partial way to specify compile time macros.

Iff there is a file with the same basename and an extension .sch
the latter is loaded into gambit via -e option.

Caveat:

lambdanative does not check the .sch file for modifications like it
does with the .scm files.  Updates to the file do not trigger a
recompile.

A real solution would be to support r7rs style library declarations.