GAMBIT: improve IPv6 fix handling client side too now
Note: This is the fix of the actual source only.

Actually building gambit with this patch does NOT already use all of
the fix.  The file _io.scm needs to be precompiled.

In order to not break the workflow when updating lambdanative and keep
this patch readable, there will be another one submitting the
precompiled _io.c file.  Both patches should be either both or none of
them applied.