TTF: do not keep global reference to potential garbage
As far as I can tell there is no need to keep the raw reference to
compiled the u8vector of a texture/font once it's been processed.
Doing so the way it was done will install a global variable holding on
the garbage thus increasing the memory consumption for no good reason.