Writeup: Detailed rationale documenting a fix.
Issue: typing a character not in the keyboard in use results in a fatal crash.

This writeup will document in unusual detail the steps and thoughts tracking it down.

It's more-or-less live - no reset changes.  Normally I would only produce the final result.
And likely I would never go into this level of details.

So bear with me: I'm prone to err as everyone else.  Beware a mess ahead.