GLGUI: fix chat widget update.
Until now the chat widget caches a pre-processed version of the
list of entries to display keyed on the `length` of the list.

Avoiding some exponential complexity during drawing, there was and
upper bound of the length maintained.  This broke once the list grew
beyond that.

This change slightly modifies the cache - and therefore renames it
within the widget.

The new caching tests ONLY whether or not the old list is the SAME in
terms of `eq?` as the new one.  It INTENSIONALY does NOT try to detect
updates to the components of the list.  This decision was made because
drawing happens much, much more frequent than updates.  It is trivial
on the update side to avoid side effects, but costly on the drawing
side to detect them.