ANDROID: try harder to handle onStop and onPause
Note: This is at the moment meant for evaluation.  Works for me, can't
break it.

However if this is going to be the final call, those file names
ANDROID_java_oncreate are no longer telling exactly where the code
ends up.

That's bad, bnut renaming them is likely even worse.