ANDROID: comment out all calls to DeleteLocalRef
I'm sorry to have introduced these in the first place.

Some documentation suggested that we really want those to be
released.  Some other documentation suggest that there is no need to
do so unless certain - rare - conditions are met: many allocated from
within a single native function.

In  practice it seems to be the case that eager use of DeleteLocalRef
leads to stuck applications.

This is NOT the final saying at this point.  It just seems to work
better without.

Edited to add: this "fix" does NOT actually fix the rot cause of the
investigation.  Looks more like mitigation.  Still we see apps getting
stuck!